requirements.txt
langchain
langchain-openai
sentence_transformers
langchain-core
langchain_community
docx2txt
pypdf
langchain_chroma
python-multipart
python-dotenv
fastapi
uvicorn

streamlit
